+++ lua/typewriter/init.lua
@@ -1,3 +1,3 @@
-require("typewriter.options")
-require("typewriter.keymaps")
-require("typewriter.lazy")
+require("typewriter.config.options")
+require("typewriter.config.keymaps")
+require("typewriter.config.lazy")