+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Refactor
 
 - Rename `me` lua module to `typewriter`. No functionality is affected.
+- Move `keymaps`, `options`, `lazy` modules in `config. No functionality is affected.
 
 ## [v0.5.0] - 2024-11-21